% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateGSVAscores.R
\name{aggregateGSVAscores}
\alias{aggregateGSVAscores}
\title{Take difference of two paired GSVA scores}
\usage{
aggregateGSVAscores(aggList, K2res)
}
\arguments{
\item{aggList}{A list where each item is a vector of 3 items:
the new name, the name of the 'up' gene set, and the name of the 'down'
gene set.}

\item{K2res}{An object of class K2. The output of runDGE_mods().}
}
\value{
An object of class K2.
}
\description{
Replaces GSVA results from paired up- and down- gene sets with the difference of
the up-regulated genes and down-regulated genes
}
\examples{
aggregateGSVAscores(aggList, K2res)
}
\keyword{clustering}
