% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cKmeans_wrapper_subsample.R
\name{cKmeans_wrapper_subsample}
\alias{cKmeans_wrapper_subsample}
\title{Wrapper for hierarchical clustering}
\usage{
cKmeans_wrapper_subsample(dataMatrix, clustList)
}
\arguments{
\item{dataMatrix}{An P x N numeric matrix of data}

\item{clustList}{List of objects to use for clustering procedure.}
}
\value{
A character string of concatenated 1's and 2's pertaining to the cluster assignment of each column in dataMatrix.
}
\description{
This fuction is a wrapper for the constrained Kmeans algorithm using 
`lcvqe` from the `conclust` package. This function will subset each
cohort down to that with the smallest number of observations.
}
\examples{
cKmeans_wrapper_subsample(dataMatrix, clusList = list())

}
\keyword{clustering}
