% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cKmeansDownsampleSqrt.R
\name{cKmeansDownsampleSqrt}
\alias{cKmeansDownsampleSqrt}
\title{Wrapper for constrained K-means on data subsampled to the square root of the 
number of observations in each cohort.}
\usage{
cKmeansDownsampleSqrt(labels, features, K2res)
}
\arguments{
\item{labels}{Vector of cohort values}

\item{features}{List of features (genes) to include}

\item{K2res}{A K2 class object.}
}
\value{
A character string of concatenated 1's and 2's pertaining to the
cohort assignments.
}
\description{
This function is a wrapper for the constrained Kmeans algorithm using
lcvqe() from the conclust package. This function will subset each
cohort down to that with the smallest number of observations.This
function is not meant to be run individually, but as a 'clustFunc'
argument for running K2tax().
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{cKm}{K2Taxonomer}
}
