% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cKmeansWrapper.R
\name{cKmeansWrapper}
\alias{cKmeansWrapper}
\title{Wrapper for constrained K-means}
\usage{
cKmeansWrapper(dataMatrix, clustList)
}
\arguments{
\item{dataMatrix}{An P x N numeric matrix of data}

\item{clustList}{List of objects to use for clustering procedure.}
}
\value{
A character string of concatenated 1's and 2's pertaining to the
cluster assignment of each column in dataMatrix.
}
\description{
This fuction is a wrapper for the constrained Kmeans algorithm using
`lcvqe` from the `conclust` package. This function is not meant to 
be run individually, but as a 'clustFunc' argument for running 
`K2preproc()`, `runK2Taxonomer()`, and `K2tax()`.
}
\examples{

dat <- scRNAseq::ReprocessedAllenData(assays = "rsem_tpm")[seq_len(50),]

eSet <- ExpressionSet(assayData = assay(dat))
pData(eSet) <- as.data.frame(colData(dat))
exprs(eSet) <- log2(exprs(eSet) + 1)

## Subset for fewer cluster labels for this example
eSet <- eSet[, !is.na(eSet$Primary.Type) & 
                 eSet$Primary.Type \%in\% c("L4 Arf5", 
                     "L4 Ctxn3", "L4 Scnn1a", "L5 Ucma", "L5a Batf3")]

## Create cell type variable with spaces
eSet$celltype <- gsub(" ", "_", eSet$Primary.Type)

## Create clustList
cL <- list(
    eMat = exprs(eSet),
    labs = eSet$celltype,
    maxIter = 10
)

## Run K2preproc to generate generate data matrix
## with a column for each celltype.
K2res <- K2preproc(eSet,
                   cohorts = "celltype",
                   featMetric = "F",
                   logCounts = TRUE)
dm <- K2data(K2res)

## Generate K=2 split with constrained K-means
cKmeansWrapperSubsample(dm, cL)

}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{cKm}{K2Taxonomer}
}
\keyword{clustering}
