% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutreeK2dendro.R
\name{cutreeK2dendro}
\alias{cutreeK2dendro}
\title{Cut K2dendro output based on specified number of clusters}
\usage{
cutreeK2dendro(K2dendrogram, k = 2)
}
\arguments{
\item{K2dendrogram}{The output of K2dendr}

\item{k}{Specified number of clusters}
}
\value{
A vector of cluster assignments
}
\description{
Returns cluster for each observation based on input k value
}
\examples{
cutreeK2dendro(K2dendrogram, k = 2)
}
\keyword{clustering}
