% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDGEhits.R
\name{getDGEhits}
\alias{getDGEhits}
\title{Extract gene set hits from K2 object}
\usage{
getDGEhits(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of runDGEmods().
getDGEhits(K2res)}
}
\description{
Creates names list of genes found in gene sets for different results.
}
