% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDGETable.R
\name{getDGETable}
\alias{getDGETable}
\title{Extract table of differential results from 'K2' object}
\usage{
getDGETable(K2res)
}
\arguments{
\item{K2res}{An object of class K2.}
}
\value{
A data.frame object.
}
\description{
Create table differential analysis results from 'K2' object.
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
               stabThresh = 0.5)

## Run differential analysis on each partition
K2res <- runDGEmods(K2res)

## Extract table of differential results
head(getDGETable(K2res))

}
\keyword{clustering}
