% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentInter.R
\name{getEnrichmentInter}
\alias{getEnrichmentInter}
\title{Create interactive table of gene set enrichment results from 'K2' object}
\usage{
getEnrichmentInter(
  K2res,
  maxFDR_score = 0.01,
  minDiff_score = NULL,
  maxPval_score = NULL,
  maxFDR_fisher = NULL,
  maxPval_fisher = NULL,
  gsNames = NULL,
  nodes = NULL
)
}
\arguments{
\item{K2res}{A K2 class object.}

\item{maxFDR_score}{Numeric. A value between 0 and 1 indicating the FDR cutoff for differential analysis of enrichment scores.}

\item{minDiff_score}{Numeric. A value between 0 and 1 indicating the mean difference cutoff for differential analysis of enrichment scores.}

\item{maxPval_score}{Numeric. A value between 0 and 1 indicating the p-value cutoff for differential analysis of enrichment scores.}

\item{maxFDR_fisher}{Numeric. A value between 0 and 1 indicating the FDR cutoff for Fisher-based overrepresentation analysis.}

\item{maxPval_fisher}{Numeric. A value between 0 and 1 indicating the p-value cutoff for Fisher-based overrepresentation analysis.}

\item{gsNames}{Character. A vector of gene sets identifiers to display.}

\item{nodes}{Character. A vector of node identifiers to display.}
}
\value{
An interactive data frame with the following columns:
\itemize{
 \item{Gene Set: }{The identifier of the gene set.}
 \item{Node: }{The identifier of the partition}
 \item{Edge: }{The identifier of partition subgroup}
 \item{Direction: }{The direction of coefficient for the assigned gene sets}
 \item{P Value Fisher: }{The p-value estimated by Fisher-based overrepresentation analysis}
 \item{FDR Fisher: }{The multiple hypothesis corrected FDR Fisher p-value, adjusted across
 all partitions}
 \item{N Overlap: }{The intersection of differentially expressed genes and gene set}
 \item{N Sig. Genes: }{The number differentially expressed genes}
 \item{N Gene Set: }{The number genes comprising the gene set}
 \item{P Value Score: }{The p-value estimated by differential analysis of enrichment scores}
 \item{FDR Score: }{The multiple hypothesis corrected FDR p-value estimated by differential analysis of enrichment scores, adjusted across
 all partitions}
 \item{Diff Score: }{The difference between the mean enrichment score of each subgroup at
 a given partition}
}
}
\description{
Create table gene set enrichment results from 'K2' object.
}
