% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenePathways.R
\name{getGenePathways}
\alias{getGenePathways}
\title{Map features (genes) to feature list (genesets)}
\usage{
getGenePathways(genesets)
}
\arguments{
\item{genesets}{A named list of features in row names of dataMatrix.}
}
\value{
A named vector
}
\description{
This function will create a named vector of collapsed names of named genesets list, separated by "; "
}
\examples{
getGenePathways(genesets)

}
\keyword{clustering}
