% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTestsModTable.R
\name{getTestsModTable}
\alias{getTestsModTable}
\title{Extract table of phenotypic variable tests from 'K2' object}
\usage{
getTestsModTable(K2res)
}
\arguments{
\item{K2res}{An object of class K2 or K2results().}
}
\value{
A data.frame object with the following columns:
\itemize{
 \item{value: }{The variable being tested}
 \item{node: }{The partition label}
 \item{edge: }{The subgroup for the given partition}
 \item{pval: }{Nominal p-value of test}
 \item{fdr: }{Benjamini-Hochberg FDR corrected p-value}
 \item{df: }{Degrees of freedom of test}
 \item{stat: }{Test statistic}
 \item{obsMean: }{Mean value across partition members}
 \item{altMean: }{Mean value for all other observations}
 \item{diffMean: }{Difference is mean}
 \item{nhits: }{The number of second label values in subgroup}
 \item{ncase: }{The total second-level label value}
 \item{nalt: }{The total first-level label value}
 \item{ndrawn: }{The total members in the subgroup}
 \item{hits: }{Members of subgroup with second-level label value}
 \item{test: }{The statistical test that produced this result}
}
}
\description{
Create table phenotypic variable results from 'K2' object.
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
            stabThresh=0.5)

infoClassVector <- c(sex='factor', score='numeric1')

K2res <- runTestsMods(K2res, infoClass=infoClassVector)

head(getTestsModTable(K2res))

}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{bh}{K2Taxonomer}
}
\keyword{clustering}
