% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_wrapper.R
\name{hclust_wrapper}
\alias{hclust_wrapper}
\title{Wrapper for hierarchical clustering}
\usage{
hclust_wrapper(dataMatrix, distMetric = "euclidean",
  aggMethod = "ward.D2")
}
\arguments{
\item{dataMatrix}{An P x N numeric matrix of data}

\item{distMetric}{A character string. One of the "method" arguments in the dist() function}

\item{aggMethod}{A character string.  One of the "method" arguments in the hclust() function}
}
\value{
A character string of concatenated 1's and 2's pertaining to the cluster assignment of each column in dataMatrix.
}
\description{
This fuction is a wrapper for the hclust, dist, and cutree functions. It outputs a string of a concatenated vector of 1's and 2's indexed by column.
}
\examples{
hclust_wrapper(dataMatrix, distMetric = "euclidean", aggMethod = "ward.D2")

}
\keyword{clustering}
