% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustWrapper.R
\name{hclustWrapper}
\alias{hclustWrapper}
\title{Wrapper for hierarchical clustering}
\usage{
hclustWrapper(labels, features, K2res)
}
\arguments{
\item{labels}{Vector of cohort values}

\item{features}{List of features (genes) to include}

\item{K2res}{A K2 class object.}
}
\value{
A character string of concatenated 1's and 2's pertaining to the
cohort assignments.
}
\description{
This fuction is a wrapper for the hclust, dist, and cutree functions. It
outputs a string of a concatenated vector of 1's and 2's indexed by column.
This function is not meant to be run individually, but as a 'clustFunc'
argument for running K2tax().
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
}
\keyword{clustering}
