% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{K2-class}
\alias{K2-class}
\title{K2 object}
\description{
This class is the output of runK2Taxonomer().
}
\section{Slots}{

\describe{
\item{\code{dataMatrix}}{An P x N numeric matrix of data}

\item{\code{info}}{A data frame with rownames that match column names in dataMatrix}

\item{\code{genesets}}{A named list of features in row names of dataMatrix}

\item{\code{gene2Pathway}}{A vector of collapsed genesets names, mapping features to
genesets}

\item{\code{eSet}}{An Expression Set object}

\item{\code{gSet}}{An Expression Set object}

\item{\code{meta}}{A named list of K2 Taxonomer information}

\item{\code{geneURL}}{A named vector of gene URLs}

\item{\code{genesetURL}}{A named vector of geneset URLs}
}}

\keyword{clustering}
