% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2dashboard.R
\name{K2dashboard}
\alias{K2dashboard}
\title{Generate interactive dashboard of K2 Taxonomer results}
\usage{
K2dashboard(
  K2res,
  analysis_name = "K2 Taxonomer",
  about = "<u>K2 Taxonomer Partitioning Results</u>",
  output_dir = "."
)
}
\arguments{
\item{K2res}{Object of class K2.}

\item{analysis_name}{Character string of the name of analysis to write files
and generate title of report.}

\item{about}{Character string of HTML code for including in the 
'about' window of the dashboard.}

\item{output_dir}{Base directory to put report.}
}
\value{
Nothing.  Writes files to a specified directory.
}
\description{
This function will generate an interactive dashboard of the annotated K=2
clustering results in a specified directory.
}
\examples{
K2dashboard(K2results, analysis_name = 'K2Taxonomer', output_dir = '.')

}
\keyword{clustering}
