% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2eMat}
\alias{K2eMat}
\alias{K2eMat<-}
\alias{K2eMat,K2-method}
\alias{K2eMat<-,K2-method}
\title{Expression matrix object used in partitioning}
\usage{
K2eMat(K2res)

K2eMat(K2res) <- value

\S4method{K2eMat}{K2}(K2res)

\S4method{K2eMat}{K2}(K2res) <- value
}
\arguments{
\item{K2res}{K2 class object.}

\item{value}{Matrix.}
}
\value{
Matrix.
}
\description{
Retrieve or assign Matrix object.
}
\examples{
data(K2res)
K2eMat(K2res)

}
