% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2eSet}
\alias{K2eSet}
\title{ExpressionSet object used in K2Taxonomer run}
\usage{
K2eSet(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
ExpressionSet.
}
\description{
Retrieve or assign ExpressionSet object.
}
\examples{
data(K2res)
K2eSet(K2res)

}
