% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2gene2Pathway}
\alias{K2gene2Pathway}
\title{Vector of collapsed pathways for which each gene belongs}
\usage{
K2gene2Pathway(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
Named vector. A vector with collapse pathway names for each gene.
}
\description{
Retrieve or assign named vector of pathways for each gene.
}
\examples{
data(K2res)
head(K2gene2Pathway(K2res))

}
