% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2genesetURL}
\alias{K2genesetURL}
\title{Assign URLs to genesets included in K2Taxonomer run}
\usage{
K2genesetURL(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
Named vector. A vector with a URL string for each geneset.
}
\description{
Retrieve or assign URLs to geneset pages.
}
\examples{
K2genesetURL(K2res)
}
