% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2gMat}
\alias{K2gMat}
\alias{K2gMat<-}
\alias{K2gMat,K2-method}
\alias{K2gMat<-,K2-method}
\title{Numeric matrix of enrichment scores}
\usage{
K2gMat(K2res)

K2gMat(K2res) <- value

\S4method{K2gMat}{K2}(K2res)

\S4method{K2gMat}{K2}(K2res) <- value
}
\arguments{
\item{K2res}{K2 class object.}

\item{value}{Matrix.}
}
\value{
Matrix.
}
\description{
Retrieve or assign matrix of enrichment scores.
}
