% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2meta}
\alias{K2meta}
\title{Meta data defining parameters of K2Taxonomer run}
\usage{
K2meta(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
Named list. Parameters used in K2Taxonomer run.
}
\description{
Retrieve or assign meta data variables.
}
\examples{
K2meta(K2res)
}
