% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{K2res}
\alias{K2res}
\title{Example K2Taxonomer output}
\format{
An object of class `K2`
}
\source{
data('sample.ExpressionSet', package='Biobase')
}
\usage{
K2res
}
\description{
An object of class, `K2`, generated by running K2Taxonomer on
`data(sample.ExpressionSet)` from `Biobase` R package.
}
\keyword{datasets}
