% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2visNetwork.R
\name{K2visNetwork}
\alias{K2visNetwork}
\title{Interactive K2 dendrogram}
\usage{
K2visNetwork(
  K2res,
  annot = c("partition", "pathways", "genes"),
  alpha = 0.05,
  limit = 5
)
}
\arguments{
\item{K2res}{A list object. The output of runK2tax().}

\item{annot}{One of 'partition', 'pathways', 'genes, to specify annotation type.}

\item{alpha}{Significance level with which to filter pathways/genes.}

\item{limit}{Number of pathways/genes to include in output.}

\item{labelsize}{Size of labels displayed in window.}
}
\value{
An interactive dendrogram created by visNetwork::visNetwork().
}
\description{
Create an interactive dendrogram of the K2 Taxonomer results
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
}
