% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcvqe.R
\name{lcvqe}
\alias{lcvqe}
\title{LCVQE algorithm.}
\usage{
lcvqe(data, k, mustLink, cantLink, maxIter = 10)
}
\arguments{
\item{data}{The unlabeled dataset}

\item{k}{Number of clusters}

\item{mustLink}{A list of must-link constraints}

\item{cantLink}{A list of cannot-link constraints}

\item{maxIter}{Number of iterations}
}
\value{
A vector that represents the labels (clusters) of the data points.
}
\description{
This function takes an unlabeled dataset and two lists of must-link and 
cannot-link constraints as input and produce a clustering as output. This 
function was originally published in the 'conclust' package,
https://github.com/cran/conclust, and was archived on CRAN on
2025-05-21, https://CRAN.R-project.org/package=conclust.
}
\details{
This algorithm finds a clustering that satisfies as many constraints as 
possible
}
\references{
\insertRef{cKm}{K2Taxonomer}
}
