% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{matrixORdgCMatrix-class}
\alias{matrixORdgCMatrix-class}
\title{K2 object}
\description{
This class is the output of K2preproc().
}
\section{Slots}{

\describe{
\item{\code{eMat}}{A numeric matrix comprising expression data used for partitioning}

\item{\code{eMatDS}}{A numeric matrix comprising expression data used for differential expression}

\item{\code{gMat}}{A numeric matrix comprising the output of runGeneSetScores().}

\item{\code{results}}{A named list object containing results from each partition}

\item{\code{colData}}{A data frame containing observation level data}

\item{\code{meta}}{A named list containing different parameters used in workflow}

\item{\code{genesets}}{A named list containing gene sets}

\item{\code{gene2Pathway}}{Vector of collapsed gene set for which each gene belongs}

\item{\code{dataMatrix}}{A numeric matrix used for different partitioning tasks}

\item{\code{geneURL}}{A named list containing URLs for each gene}

\item{\code{genesetURL}}{A named list containing URLs for each gene set}
}}

\keyword{clustering}
