% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenePathway.R
\name{plotGenePathway}
\alias{plotGenePathway}
\title{Plot gene expression or enrichment scores}
\usage{
plotGenePathway(
  K2res,
  feature,
  node,
  type = c("eMatDS", "eMat", "gMat"),
  use_plotly = TRUE,
  subsample = TRUE
)
}
\arguments{
\item{K2res}{A K2 class object.}

\item{node}{Character. A specific node identifier in K2res object.}

\item{type}{Character. One of 'eMatDS' (default), 'eMat', or 'gMat'.}

\item{use_plotly}{Logical. Whether to create a static plot, FALSE, or
interactive plot, TRUE (default).}

\item{subsample}{Logical. Whether to downsample the output object in order to
save memory. This minimal impact on plot visual. Default is TRUE.}

\item{features}{Character. Either a gene or gene set identifier.}
}
\value{
Either a ggplot or ggplotly object.
}
\description{
This function creates a plot for visualizing gene expression between
partitions, either as a static plot or interactive with plotly.
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
}
