% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDGE_mods.R
\name{runDGE_mods}
\alias{runDGE_mods}
\title{Perform differential analysis of gene splits}
\usage{
runDGE_mods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds limma differential analysis results to the output of K2tax().
}
\examples{
runDGE_mods(K2res)
}
\keyword{clustering}
