% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDGEmods.R
\name{runDGEmods}
\alias{runDGEmods}
\title{Perform differential analysis of gene splits}
\usage{
runDGEmods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds limma differential analysis results to the output of K2tax().
}
\examples{
runDGEmods(K2res)
}
\keyword{clustering}
