% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDSSE_mods.R
\name{runDSSE_mods}
\alias{runDSSE_mods}
\title{Perform differential analysis of single-sample gene set enrichment}
\usage{
runDSSE_mods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds limma differential analysis results of single-sample enrichment scores to to the output of K2tax().
}
\examples{
runDSSE_mods(K2res)
}
\keyword{clustering}
