% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDSSEmods.R
\name{runDSSEmods}
\alias{runDSSEmods}
\title{Perform differential analysis of single-sample gene set enrichment}
\usage{
runDSSEmods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds limma differential analysis results of single-sample enrichment scores
to to the output of K2tax().
}
\examples{
runDSSEmods(K2res)
}
\keyword{clustering}
