% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSE_mods.R
\name{runGSE_mods}
\alias{runGSE_mods}
\title{Perform hyperenrichment of split specific gene signatures}
\usage{
runGSE_mods(K2res, genesets = NULL, qthresh = NULL, cthresh = NULL,
  ntotal = NULL)
}
\arguments{
\item{K2res}{An object of class K2. The output of runDGE_mods().}

\item{genesets}{A named list of feature IDs}

\item{qthresh}{A numeric value between 0 and 1 of the FDR cuttoff to define feature sets.}

\item{cthresh}{A positive value for the coefficient cuttoff to define feature sets.}
}
\value{
An object of class K2.
}
\description{
Adds hyperenrichment analysis results to the output of runDGE_mods().
}
\examples{
runGSE_mods(K2res)
}
\keyword{clustering}
