% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSVA_mods.R
\name{runGSVA_mods}
\alias{runGSVA_mods}
\title{Perform ssGSVA on gene sets and get set signature hits}
\usage{
runGSVA_mods(K2res, ssGSEAalg = NULL, ssGSEAcores = NULL, ...)
}
\arguments{
\item{K2res}{An object of class K2. The output of runDGE_mods().}

\item{...}{Additional arguments passed onto GSVA::gsva()}

\item{genesets}{A named list of feature IDs}

\item{qthresh}{A numeric value between 0 and 1 of the FDR cuttoff to define feature sets.}

\item{cthresh}{A positive value for the coefficient cuttoff to define feature sets.}
}
\value{
An object of class K2.
}
\description{
Adds hyperenrichment analysis results to the output of runDGE_mods().
}
\examples{
runGSVA_mods(K2res)
}
\keyword{clustering}
