% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSVAmods.R
\name{runGSVAmods}
\alias{runGSVAmods}
\title{Perform ssGSVA on gene sets and get set signature hits}
\usage{
runGSVAmods(K2res, ssGSEAalg = NULL, ssGSEAcores = NULL, ...)
}
\arguments{
\item{K2res}{An object of class K2. The output of runDGEmods().}

\item{...}{Additional arguments passed onto GSVA::gsva()}

\item{genesets}{A named list of feature IDs}

\item{qthresh}{A numeric value between 0 and 1 of the FDR cuttoff to define
feature sets.}

\item{cthresh}{A positive value for the coefficient cuttoff to define
feature sets.}
}
\value{
An object of class K2.
}
\description{
Adds hyperenrichment analysis results to the output of runDGEmods().
}
\examples{
runGSVAmods(K2res)
}
\keyword{clustering}
