% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests_mods.R
\name{runTests_mods}
\alias{runTests_mods}
\title{Perform tests to test for associations between modules and meta-variables}
\usage{
runTests_mods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds statistical tests results to the output of K2tax() 
based on numeric and factor variables in info.
}
\examples{
runTests_mods(K2res)
}
\keyword{clustering}
