% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTestsMods.R
\name{runTestsMods}
\alias{runTestsMods}
\title{Perform tests to test for associations between modules and meta-variables}
\usage{
runTestsMods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds statistical tests results to the output of K2tax()
based on numeric and factor variables in info.
}
\examples{
runTestsMods(K2res)
}
\keyword{clustering}
