% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cKmeansWrapper.R
\name{cKmeansWrapper}
\alias{cKmeansWrapper}
\title{Wrapper for constrained K-means}
\usage{
cKmeansWrapper(dataMatrix, clustList)
}
\arguments{
\item{dataMatrix}{An P x N numeric matrix of data}

\item{clustList}{List of objects to use for clustering procedure.}
}
\value{
A character string of concatenated 1's and 2's pertaining to the
cluster assignment of each column in dataMatrix.
}
\description{
This fuction is a wrapper for the constrained Kmeans algorithm using
`lcvqe` from the `conclust` package.
}
\keyword{clustering}
