% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cKmeansWrapperSubsample.R
\name{cKmeansWrapperSubsample}
\alias{cKmeansWrapperSubsample}
\title{Wrapper for constrained K-means on subsampled data}
\usage{
cKmeansWrapperSubsample(dataMatrix, clustList)
}
\arguments{
\item{dataMatrix}{An P x N numeric matrix of data}

\item{clustList}{List of objects to use for clustering procedure.}
}
\value{
A character string of concatenated 1's and 2's pertaining to the
cluster assignment of each column in dataMatrix.
}
\description{
This fuction is a wrapper for the constrained Kmeans algorithm using
`lcvqe` from the `conclust` package. This function will subset each
cohort down to that with the smallest number of observations.
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{cKm}{K2Taxonomer}
}
\keyword{clustering}
