% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDGETable.R
\name{getDGETable}
\alias{getDGETable}
\title{Extract table of differential results from 'K2' object}
\usage{
getDGETable(K2res)
}
\arguments{
\item{K2res}{A K2 class object.}
}
\value{
A data.frame object with the following columns:
\itemize{
 \item{gene: }{The identifier of the gene}
 \item{coef: }{The difference between the means of each subgroup at
 a given partition}
 \item{mean: }{The mean across all observations at the given
 partition}
 \item{t: }{The test statistic estimated by differential analysis}
 \item{pval: }{The p-value estimated by differential analysis}
 \item{fdr: }{The multiple hypothesis corrected fdr p-value, adjusted across
 all partitions}
 \item{edge: }{Indication of which subgroup the gene was assigned at a given
 partition}
 \item{node: }{The identifier of the partition}
 \item{direction: }{The direction of coefficient for the assigned gene}
}
}
\description{
Create table differential analysis results from 'K2' object.
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{limma}{K2Taxonomer}
 \insertRef{bh}{K2Taxonomer}
}
\keyword{clustering}
