% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDGETable.R
\name{getDGETable}
\alias{getDGETable}
\title{Extract table of differential results from "K2" object}
\usage{
getDGETable(K2res)
}
\arguments{
\item{K2res}{An object of class K2 or K2results().}
}
\value{
A data.frame object.
}
\description{
Create table differential analysis results from "K2" object.
}
\examples{
getDGETable(K2res)
}
\keyword{clustering}
