% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentTable.R
\name{getEnrichmentTable}
\alias{getEnrichmentTable}
\title{Extract table of enrichment results from 'K2' object}
\usage{
getEnrichmentTable(K2res)
}
\arguments{
\item{K2res}{An object of class K2 or K2results().}
}
\value{
A data.frame object.
}
\description{
Create table hyper- and single-sample enrichment results from 'K2' object.
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
               stabThresh = 0.5)

## Run differential analysis on each partition
K2res <- runDGEmods(K2res)

## Create dummy set of gene sets
DGEtable <- getDGETable(K2res)
genes <- unique(DGEtable$gene)
genesetsMadeUp <- list(
    GS1 = genes[1:50],
    GS2 = genes[51:100],
    GS3 = genes[101:150]
)

## Run gene set hyperenrichment
K2res <- runGSEmods(K2res, 
                    genesets = genesetsMadeUp,
                    qthresh = 0.1)
                    
## Run GSVA on genesets           
K2res <- runGSVAmods(K2res, 
                     ssGSEAalg = "gsva",
                     ssGSEAcores = 1,
                     verbose = FALSE)

head(getEnrichmentTable(K2res))

}
\keyword{clustering}
