% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnrichmentTable.R
\name{getEnrichmentTable}
\alias{getEnrichmentTable}
\title{Extract table of enrichment results from 'K2' object}
\usage{
getEnrichmentTable(K2res)
}
\arguments{
\item{K2res}{An object of class K2 or K2results().}
}
\value{
A data.frame object.
}
\description{
Create table hyper- and single-sample enrichment results from 'K2' object.
}
\examples{
getEnrichementTable(K2res)
}
\keyword{clustering}
