% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclustWrapper.R
\name{hclustWrapper}
\alias{hclustWrapper}
\title{Wrapper for hierarchical clustering}
\usage{
hclustWrapper(dataMatrix, clustList)
}
\arguments{
\item{dataMatrix}{An P x N numeric matrix of data}

\item{clustList}{List of objects to use for clustering procedure.}
}
\value{
A character string of concatenated 1's and 2's pertaining to the
cluster assignment of each column in dataMatrix.
}
\description{
This fuction is a wrapper for the hclust, dist, and cutree functions. It
outputs a string of a concatenated vector of 1's and 2's indexed by column.
}
\examples{
hclustWrapper(dataMatrix, clusList = list())

}
\keyword{clustering}
