% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2dashboard.R
\name{K2dashboard}
\alias{K2dashboard}
\title{Generate interactive dashboard of K2 Taxonomer results}
\usage{
K2dashboard(
  K2results,
  analysis_name = "K2Taxonomer",
  output_dir = ".",
  compile = TRUE
)
}
\arguments{
\item{K2results}{Object of class K2. The output of runK2Taxonomer().}

\item{analysis_name}{Character string of the name of analysis to write files
and generate title of report.}

\item{output_dir}{Base directory to put report.}

\item{compile}{Logical. Whether or not to compile dashboard now.}
}
\value{
Nothing.  Writes files to a specified directory.
}
\description{
This function will generate an interactive dashboard of the annotated K=2
clustering results.
}
\examples{
K2dashboard(K2results, analysis_name = 'K2Taxonomer', output_dir = '.',
compile = TRUE)

}
\keyword{clustering}
