% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2dashboard.R
\name{K2dashboard}
\alias{K2dashboard}
\title{Generate interactive dashboard of K2 Taxonomer results}
\usage{
K2dashboard(
  K2res,
  analysis_name = "K2Taxonomer",
  about = TRUE,
  output_dir = "."
)
}
\arguments{
\item{K2res}{A K2 class object.}

\item{analysis_name}{Character string of the name of analysis to write files
and generate title of report.}

\item{about}{Logical. Whether to create an 'about' tab in the dashboard.}

\item{output_dir}{Base directory to put report.}
}
\value{
Nothing.  Writes files to a specified directory.
}
\description{
This function will generate an interactive dashboard of the annotated K2
Taxonomer results in a specified directory.
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
}
\keyword{clustering}
