% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2dendro.R
\name{K2dendro}
\alias{K2dendro}
\title{Reformat K2Taxonomer results to dendrogram object}
\usage{
K2dendro(K2res)
}
\arguments{
\item{K2res}{A K2 class object.}
}
\value{
An object of class dendrogram.
}
\description{
Reformats the output of K2tax() to an object of class, dendrogram,
which can be easily plotted.
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
}
\keyword{clustering}
