% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2eSet}
\alias{K2eSet}
\alias{K2eSet<-}
\alias{K2eSet,K2-method}
\alias{K2eSet<-,K2-method}
\title{ExpressionSet object used in K2Taxonomer run}
\usage{
K2eSet(K2res)

K2eSet(K2res) <- value

\S4method{K2eSet}{K2}(K2res)

\S4method{K2eSet}{K2}(K2res) <- value
}
\arguments{
\item{K2res}{K2 class object.}

\item{value}{ExpressionSet.}
}
\value{
ExpressionSet.
}
\description{
Retrieve or assign ExpressionSet object.
}
\examples{
data(K2res)
K2eSet(K2res)

}
