% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2geneURL}
\alias{K2geneURL}
\title{Assign URLs to genes included in K2Taxonomer run}
\usage{
K2geneURL(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
Named vector. A vector with a URL string for each gene.
}
\description{
Retrieve or assign URLs to gene pages.
}
\examples{
data(K2res)
K2geneURL(K2res)

}
