% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2gSet}
\alias{K2gSet}
\title{ExpressionSet object of of GSVA output}
\usage{
K2gSet(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
ExpressionSet.
}
\description{
Retrieve or assign ExpressionSet object of GSVA output.
}
\examples{
data(K2res)
K2gSet(K2res)

}
