% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2info}
\alias{K2info}
\title{Data frame of phenotypic information from K2 object}
\usage{
K2info(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
Data frame. Observation information from K2Taxonomer run.
}
\description{
Retrieve or assign observation information in K2Taxonomer run.
}
\examples{
data(K2res)
head(K2info(K2res))

}
