% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2preproc.R
\name{K2preproc}
\alias{K2preproc}
\title{Function to create K2 object for pre-processing}
\usage{
K2preproc(eSet, cohorts = NULL, vehicle = NULL, covariates = NULL,
  use = c("Z", "MEAN"), nFeats = nrow(eSet) * 0.02,
  featMetric = c("square", "mad"), nBoots = 500,
  clustFunc = hclust_wrapper, linkage = "mcquitty", info = NULL,
  infoClass = NULL, genesets = NULL, qthresh = 0.05, cthresh = 0,
  ntotal = 20000, ssGSEAalg = c("ssgsea", "gsva", "zscore", "plage"),
  ssGSEAminQvalue = 0.01, ssGSEAcores = 0, oneoff = TRUE,
  stabThresh = 0, geneURL = NULL, genesetURL = NULL)
}
\arguments{
\item{eSet}{An expression set object.}

\item{cohorts}{The column in phenotype data of eSet that has cohort ID's. Default NULL if no pre-processing of data.}

\item{vehicle}{The value in the cohort variable that contains the vehicle ID. Default NULL if no vehicle to be used.}

\item{covariates}{Covariates in phenotype data of eSet to control for in differential analysis.}

\item{use}{Generate test statistics or use means from differential analysis for clustering.}

\item{nFeats}{A numeric value <= P of subsets of the data to use.}

\item{nBoots}{A numeric value of the number of bootstraps to run at each split.}

\item{clustFunc}{Wrapper function to cluster a P x N (See details).}

\item{linkage}{Linkage criteria for splitting cosine matrix ("method" in hclust). "average" by default.}

\item{info}{A data frame with rownames that match column names in dataMatrix.}

\item{infoClass}{= A named vector denoted types of tests to run on metavariables.}

\item{genesets}{A named list of features in row names of dataMatrix.}

\item{qthresh}{A numeric value between 0 and 1 of the FDR cuttoff to define feature sets.}

\item{cthresh}{A positive value for the coefficient cuttoff to define feature sets.}

\item{oneoff}{Logical. Allow 1 member clusters?}

\item{stabThresh}{Threshold for ending clustering.}

\item{geneURL}{Optional. Named list of URLs to gene information.}

\item{genesetURL}{Optional. Named list of URLs to geneset information.}

\item{featMatric}{Metric to use to assign variance/signal score. Options are
"square" (default) use square values and "mad" to use MAD scores.}
}
\description{
This function will generate an object of class, K2.  This will run pre-processing functions for running K2 Taxonomer procedure.
}
\examples{
K2preproc(eSet)
}
\keyword{clustering}
