% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{K2results}
\alias{K2results}
\title{Annotated list of partition assignments from K2Taxonomer run}
\usage{
K2results(K2res)
}
\arguments{
\item{K2res}{K2 class object.}
}
\value{
List. A list with an element for each partition.
}
\description{
Retrieve or assign annotated list of subgroups from K2Taxonomer
run.
}
\examples{
data(K2res)
K2resList <- K2results(K2res)

}
