% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2tax.R
\name{K2tax}
\alias{K2tax}
\title{Function K2 Taxonomer}
\usage{
K2tax(K2res, nFeats = NULL, featMetric = NULL, nBoots = NULL,
  clustFunc = NULL, clustCors = NULL, clustList = NULL,
  linkage = NULL, oneoff = NULL, stabThresh = NULL)
}
\arguments{
\item{nFeats}{A numeric value <= P of subsets of the data to use.}

\item{nBoots}{A numeric value of the number of bootstraps to run at each split.}

\item{clustFunc}{Wrapper function to cluster a P x N (See details).}

\item{clustCors}{Number of cores to use for clustering.}

\item{clustList}{List of objects to use for clustering procedure.}

\item{linkage}{Linkage criteria for splitting cosine matrix ("method" in hclust).}

\item{oneoff}{Logical. Allow 1 member clusters?}

\item{stabThresh}{A numeric value < 1, to set stopping threshold (use any negative value for no threshold).}

\item{featMatric}{Metric to use to assign variance/signal score. Options are
"square" (default), "mad" to use MAD scores, "sd" to use standard deviation}
}
\value{
An object of class K2.
}
\description{
This function performs K2 Taxonomer procedure only. Arguments values are extracted from K2meta(K2res) unless othewise specified.
}
\examples{
K2tax(K2res)

}
\keyword{clustering}
