% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K2visNetwork.R
\name{K2visNetwork}
\alias{K2visNetwork}
\title{Interactive K2 dendrogram}
\usage{
K2visNetwork(K2res, labelsize = 50)
}
\arguments{
\item{K2res}{A K2 class object.}

\item{labelsize}{Size of labels displayed in window.}
}
\value{
An interactive dendrogram created by visNetwork::visNetwork().
}
\description{
Create an interactive dendrogram of the K2 Taxonomer results
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
}
