% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDGEmods.R
\name{runDGEmods}
\alias{runDGEmods}
\title{Perform differential analysis of gene splits}
\usage{
runDGEmods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds limma differential analysis results to the output of K2tax().
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
            stabThresh=0.5)

## Run differential analysis on each partition
K2res <- runDGEmods(K2res)

}
\references{
\insertRef{reed_2020}{K2Taxonomer}
    \insertRef{limma}{K2Taxonomer}
    \insertRef{bh}{K2Taxonomer}
}
\keyword{clustering}
