% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDGEmods.R
\name{runDGEmods}
\alias{runDGEmods}
\title{Perform differential analysis between subgroups at each partition}
\usage{
runDGEmods(
  K2res,
  DGEmethod = NULL,
  DGEexpThreshold = NULL,
  vehicle = NULL,
  variables = NULL,
  block = NULL,
  logCounts = NULL
)
}
\arguments{
\item{DGEmethod}{Character. Method for running differential gene expression analyses. Use one of either 'limma' (default) or 'mast'.}

\item{DGEexpThreshold}{Numeric. A value between 0 and 1 indicating for filtering lowly expressed genes for partition-specific differential gene expression. Proportion of observations with counts > 0 in at least one subgroup at a specific partition.}

\item{vehicle}{The value in the cohort variable that contains the ID of observation to use as control. Default NULL if no vehicle to be used.}

\item{variables}{Character. Columns in meta data of 'object' to control for in differential analyses.}

\item{block}{Character. Block parameter in limma for modeling higherarchical data structure, such as multiple observations per individual.}

\item{logCounts}{Logical. Whether or not expression values are log-scale counts or log normalized counts from RNA-seq. Default is TRUE.}
}
\value{
An object of class K2.
}
\description{
Adds differential analysis results to the output of K2tax().
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{limma}{K2Taxonomer}
 \insertRef{bh}{K2Taxonomer}
}
\keyword{clustering}
