% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDSSEmods.R
\name{runDSSEmods}
\alias{runDSSEmods}
\title{Perform differential analysis of enrichment scores between subgroups at each
partition}
\usage{
runDSSEmods(K2res, vehicle = NULL, variables = NULL, block = NULL)
}
\arguments{
\item{vehicle}{The value in the cohort variable that contains the ID of observation to use as control. Default NULL if no vehicle to be used.}

\item{variables}{Character. Columns in meta data of 'object' to control for in differential analyses.}

\item{block}{Character. Block parameter in limma for modeling higherarchical data structure, such as multiple observations per individual.}
}
\value{
An object of class K2.
}
\description{
Adds differential analysis results of single-sample enrichment scores
to to the output of K2tax().
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{limma}{K2Taxonomer}
 \insertRef{bh}{K2Taxonomer}
 \insertRef{gsva}{K2Taxonomer}
}
\keyword{clustering}
