% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDSSEmods.R
\name{runDSSEmods}
\alias{runDSSEmods}
\title{Perform differential analysis of single-sample gene set enrichment}
\usage{
runDSSEmods(K2res)
}
\arguments{
\item{K2res}{An object of class K2. The output of K2tax().}
}
\value{
An object of class K2.
}
\description{
Adds limma differential analysis results of single-sample enrichment scores
to to the output of K2tax().
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
               stabThresh = 0.5)

## Run differential analysis on each partition
K2res <- runDGEmods(K2res)

## Create dummy set of gene sets
DGEtable <- getDGETable(K2res)
genes <- unique(DGEtable$gene)
genesetsMadeUp <- list(
    GS1 = genes[1:50],
    GS2 = genes[51:100],
    GS3 = genes[101:150]
)

## Run gene set hyperenrichment
K2res <- runGSEmods(K2res, 
                    genesets = genesetsMadeUp,
                    qthresh = 0.1)
                    
## Run GSVA on genesets           
K2res <- runGSVAmods(K2res, 
                     ssGSEAalg = "gsva",
                     ssGSEAcores = 1,
                     verbose = FALSE)

## Run differential analysis on GSVA results
K2res <- runDSSEmods(K2res)

}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{limma}{K2Taxonomer}
 \insertRef{bh}{K2Taxonomer}
 \insertRef{gsva}{K2Taxonomer}
}
\keyword{clustering}
