% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSEmods.R
\name{runGSEmods}
\alias{runGSEmods}
\title{Perform hyperenrichment of split specific gene signatures}
\usage{
runGSEmods(
  K2res,
  genesets = NULL,
  qthresh = NULL,
  cthresh = NULL,
  ntotal = NULL
)
}
\arguments{
\item{K2res}{An object of class K2. The output of runDGEmods().}

\item{genesets}{A named list of feature IDs}

\item{qthresh}{A numeric value between 0 and 1 of the FDR cuttoff to define
feature sets.}

\item{cthresh}{A positive value for the coefficient cuttoff to define
feature sets.}
}
\value{
An object of class K2.
}
\description{
Adds hyperenrichment analysis results to the output of runDGEmods().
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
               stabThresh = 0.5)

## Run differential analysis on each partition
K2res <- runDGEmods(K2res)

## Create dummy set of gene sets
DGEtable <- getDGETable(K2res)
genes <- unique(DGEtable$gene)
genesetsMadeUp <- list(
    GS1 = genes[1:50],
    GS2 = genes[51:100],
    GS3 = genes[101:150]
)

## Run gene set hyperenrichment
K2res <- runGSEmods(K2res, 
                    genesets = genesetsMadeUp,
                    qthresh = 0.1)
                    
}
\references{
\insertRef{reed_2020}{K2Taxonomer}
 \insertRef{bh}{K2Taxonomer}
}
\keyword{clustering}
