% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSVAmods.R
\name{runGSVAmods}
\alias{runGSVAmods}
\title{Perform ssGSVA on gene sets and get set signature hits}
\usage{
runGSVAmods(K2res, ssGSEAalg = NULL, ssGSEAcores = NULL, ...)
}
\arguments{
\item{K2res}{An object of class K2. The output of runDGEmods().}

\item{ssGSEAalg}{A character string, specifying which algorithm to use for
running the gsva() function from the GSVA package. Options are 'gsva',
'ssgsea', 'zscore', and 'plage'. 'gsva' by default.}

\item{ssGSEAcores}{Number of cores to use for running gsva() from the GSVA
package. Default is 1.}

\item{...}{Additional arguments passed onto GSVA::gsva()}
}
\value{
An object of class K2.
}
\description{
Adds hyperenrichment analysis results to the output of runDGEmods().
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Pre-process and create K2 object
K2res <- K2preproc(sample.ExpressionSet)

## Run K2 Taxonomer algorithm
K2res <- K2tax(K2res,
            stabThresh=0.5)

## Run differential analysis on each partition
K2res <- runDGEmods(K2res)

## Create dummy set of gene sets
DGEtable <- getDGETable(K2res)
genes <- unique(DGEtable$gene)
genesetsMadeUp <- list(
    GS1=genes[1:50],
    GS2=genes[51:100],
    GS3=genes[101:150])

## Run gene set hyperenrichment
K2res <- runGSEmods(K2res,
                genesets=genesetsMadeUp,
                qthresh=0.1)

## Run GSVA on genesets
K2res <- runGSVAmods(K2res,
                ssGSEAalg='gsva',
                ssGSEAcores=1,
                verbose=FALSE)

}
\references{
    \insertRef{reed_2020}{K2Taxonomer}
    \insertRef{gsva}{K2Taxonomer}
}
\keyword{clustering}
