% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runK2Taxonomer.R
\name{runK2Taxonomer}
\alias{runK2Taxonomer}
\title{Wrapper function to run K2Taxonomer algorithm and annotation}
\usage{
runK2Taxonomer(
    eSet,
    cohorts = NULL,
    vehicle = NULL,
    covariates = NULL,
    block = NULL,
    logCounts = FALSE,
    use = c("Z", "MEAN"),
    nFeats = "sqrt",
    featMetric = c("mad", "sd", "Sn", "Qn", "F", "square"),
    recalcDataMatrix = FALSE,
    nBoots = 500,
    clustFunc = hclustWrapper,
    clustCors = 1,
    clustList = list(),
    linkage = c("mcquitty", "ward.D", "ward.D2", "single", "complete",
        "average", "centroid"),
    info = NULL,
    infoClass = NULL,
    genesets = NULL,
    qthresh = 0.05,
    cthresh = 0,
    ntotal = 20000,
    ssGSEAalg = c("gsva", "ssgsea", "zscore", "plage"),
    ssGSEAcores = 1,
    oneoff = TRUE,
    stabThresh = 0,
    geneURL = NULL,
    genesetURL = NULL
)
}
\arguments{
\item{eSet}{An expression set object.}

\item{cohorts}{The column in phenotype data of eSet that has cohort ID's.
Default NULL if no pre-processing of data.}

\item{vehicle}{The value in the cohort variable that contains the vehicle
ID. Default NULL if no vehicle to be used.}

\item{covariates}{Covariates in phenotype data of eSet to control for in
differential analysis.}

\item{block}{Block parameter in limma for modelling random-like effects.}

\item{logCounts}{Logical. Whether or not expression values are log-scale
counts or log normalized counts from RNA-seq. Default is FALSE.}

\item{use}{Character string. Options are 'Z' to generate test statistics or
'MEAN' to use means from differential analysis for clustering.}

\item{nFeats}{'sqrt' or a numeric value <= number of features to subset the
features for each partition.}

\item{featMetric}{Metric to use to assign variance/signal score. Options are:
'mad' (default), 'mad', 'Sn', 'Qn', 'F', and 'square'.}

\item{recalcDataMatrix}{Logical. Recalculate dataMatrix for each partion?
Default is FALSE.}

\item{nBoots}{A numeric value of the number of bootstraps to run at each
split.}

\item{clustFunc}{Wrapper function to cluster a P x N (See details).}

\item{clustCors}{Number of cores to use for clustering.}

\item{clustList}{List of objects to use for clustering procedure.}

\item{linkage}{Linkage criteria for splitting cosine matrix ('method' in
hclust). 'average' by default.}

\item{info}{A data frame with rownames that match column names in dataMatrix.}

\item{infoClass}{A named vector denoted types of tests to run on
metavariables.}

\item{genesets}{A named list of features in row names of dataMatrix.}

\item{qthresh}{A numeric value between 0 and 1 of the FDR cuttoff to define
feature sets.}

\item{cthresh}{A positive value for the coefficient cuttoff to define
feature sets.}

\item{ntotal}{A positive value to use as the background feature count. 20000
by default.}

\item{ssGSEAalg}{A character string, specifying which algorithm to use for
running the gsva() function from the GSVA package. Options are 'gsva',
'ssgsea', 'zscore', and 'plage'. 'gsva' by default.}

\item{ssGSEAcores}{Number of cores to use for running gsva() from the GSVA
package. Default is 1.}

\item{oneoff}{Logical. Allow 1 member clusters?}

\item{stabThresh}{Threshold for ending clustering.}

\item{geneURL}{Optional. Named list of URLs to gene information.}

\item{genesetURL}{Optional. Named list of URLs to geneset information.}
}
\value{
An object of class, `K2`.
}
\description{
This function will generate an object of class, K2.  This will run the K2
Taxonomer procedure, differential analysis, and finally hyperenrichment on a
named list of feature sets.
}
\examples{
## Read in ExpressionSet object
library(Biobase)
data(sample.ExpressionSet)

## Create dummy set of gene sets
genes <- rownames(sample.ExpressionSet)
genesetsMadeUp <- list(
    GS1=genes[1:50],
    GS2=genes[51:100],
    GS3=genes[101:150])

## Run K2 Taxonomer wrapper
K2Res <- runK2Taxonomer(sample.ExpressionSet,
                    genesets=genesetsMadeUp,
                    qthresh=0.1,
                    ssGSEAalg='gsva',
                    ssGSEAcores=1,
                    stabThresh=0.5)

}
\references{
    \insertRef{reed_2020}{K2Taxonomer}
}
